delete fix and PR comments
