+++ publishing/admin.py
@@ -1,10 +1,10 @@
 from django.contrib import admin
 from django.utils.translation import ugettext_lazy as _
 
-from publishing.models.envelope import Envelope
-from publishing.models.operational_status import OperationalStatus
-from publishing.models.packaged_workbasket import PackagedWorkBasket
-from publishing.models.state import ProcessingState
+from publishing.models import Envelope
+from publishing.models import OperationalStatus
+from publishing.models import PackagedWorkBasket
+from publishing.models import ProcessingState
 
 
 class CustomProcessingStateFilter(admin.SimpleListFilter):
@@ -133,9 +133,3 @@ class OperationalStatusAdmin(admin.ModelAdmin):
 admin.site.register(OperationalStatus, OperationalStatusAdmin)
 
 admin.site.register(Envelope, EnvelopeAdmin)
-
-
-#   Envelope.objects.for_year()
-#     .filter(
-#         packagedworkbaskets__processing_state=ProcessingState.SUCCESSFULLY_PROCESSED,
-#     )
