+++ publishing/models/envelope.py
@@ -18,16 +18,25 @@ from common.models.mixins import TimestampedMixin
 from common.serializers import validate_envelope
 from exporter.serializers import MultiFileEnvelopeTransactionSerializer
 from exporter.util import dit_file_generator
-from publishing.models.exceptions import EnvelopeCurrentlyProccessing
-from publishing.models.exceptions import EnvelopeInvalidQueuePosition
-from publishing.models.exceptions import EnvelopeNoTransactions
-from publishing.models.packaged_workbasket import PackagedWorkBasket
-from publishing.models.state import ProcessingState
+from publishing.models import PackagedWorkBasket
+from publishing.models import ProcessingState
 from publishing.storages import EnvelopeStorage
 from taric import validators
 from workbaskets.models import WorkBasket
 
 logger = logging.getLogger(__name__)
+
+# Exceptions
+class EnvelopeCurrentlyProccessing(Exception):
+    pass
+
+
+class EnvelopeInvalidQueuePosition(Exception):
+    pass
+
+
+class EnvelopeNoTransactions(Exception):
+    pass
 
 
 class EnvelopeManager(Manager):
